lab-2 task-4 func return value is updated

+++ hw2/210102002019_TASK-4_func1.c
@@ -6,7 +6,7 @@
  *		Email: h.kinac2021@gtu.edu.tr
  */
 
-int func1(int number)
+unsigned long long func1(int number)
 {
     if (number == 0)
         return 1;
